requirements.txt
annotated-types==0.7.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
filetype==1.2.0
fonttools==4.60.1
idna==3.11
ImageHash==4.3.1
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
llvmlite==0.44.0
MarkupSafe==3.0.3
matplotlib==3.10.0
minify_html==0.16.4
multimethod==1.12
networkx==3.5
numba==0.61.0
numpy==2.1.3
packaging==25.0
pandas==2.3.3
patsy==1.0.1
phik==0.12.5
pillow==12.0.0
puremagic==1.30
pydantic==2.12.2
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
statsmodels==0.14.5
tqdm==4.67.1
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
visions==0.8.1
wordcloud==1.9.4
ydata-profiling==4.17.0
